+++ bindings/javascript/js-binding.d.ts
@@ -19,6 +19,10 @@ export interface Options {
   keepStyleTags?: boolean
   /** Keep "link" tags after inlining. */
   keepLinkTags?: boolean
+  /** Keep "at-rules" after inlining. */
+  keepAtRules?: boolean
+  /** Remove trailing semicolons and spaces between properties and values. */
+  minifyCss?: boolean
   /** Used for loading external stylesheets via relative URLs. */
   baseUrl?: string
   /** Whether remote stylesheets should be loaded or not. */